# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and enquiry API for the Airlink Aviation website
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ErrorResponseFields = {[key: string]: string};

export interface ErrorResponse {
  error: string;
  fields?: ErrorResponseFields;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  category: string;
  shortDescription: string;
  image: string;
  eyebrow: string;
  featured: boolean;
}

export interface Specification {
  label: string;
  value: string;
}

export type ProductDetail = Product & {
  description: string;
  features: string[];
  applications: string[];
  specifications: Specification[];
  relatedProducts: Product[];
};

export interface Resource {
  id: number;
  title: string;
  description: string;
  category: string;
  type: string;
  actionLabel: string;
  href: string;
}

export interface SiteSummary {
  productCount: number;
  capabilityCount: number;
  supportLabel: string;
  qualityLabel: string;
}

export interface EnquiryInput {
  /** @nullable */
  productSlug?: string | null;
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  company: string;
  email: string;
  /** @minLength 7 */
  phone: string;
  /** @minLength 2 */
  requirement: string;
  /** @minLength 10 */
  message: string;
}

export interface EnquiryReceipt {
  id: string;
  message: string;
}

export type GetProductsParams = {
category?: string;
search?: string;
featured?: boolean;
};

export type GetResourcesParams = {
category?: string;
};