# lib/api-zod/src/generated/types/enquiryInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and enquiry API for the Airlink Aviation website
 * OpenAPI spec version: 0.1.0
 */

export interface EnquiryInput {
  /** @nullable */
  productSlug?: string | null;
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  company: string;
  email: string;
  /** @minLength 7 */
  phone: string;
  /** @minLength 2 */
  requirement: string;
  /** @minLength 10 */
  message: string;
}