# lib/api-zod/src/generated/types/enquiryReceipt.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and enquiry API for the Airlink Aviation website
 * OpenAPI spec version: 0.1.0
 */

export interface EnquiryReceipt {
  id: string;
  message: string;
}