# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and enquiry API for the Airlink Aviation website
 * OpenAPI spec version: 0.1.0
 */
import type { ErrorResponseFields } from './errorResponseFields';

export interface ErrorResponse {
  error: string;
  fields?: ErrorResponseFields;
}