# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and enquiry API for the Airlink Aviation website
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  slug: string;
  category: string;
  shortDescription: string;
  image: string;
  eyebrow: string;
  featured: boolean;
}