# lib/api-zod/src/generated/types/productDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and enquiry API for the Airlink Aviation website
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';
import type { Specification } from './specification';

export type ProductDetail = Product & {
  description: string;
  features: string[];
  applications: string[];
  specifications: Specification[];
  relatedProducts: Product[];
};