# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and enquiry API for the Airlink Aviation website
 * OpenAPI spec version: 0.1.0
 */

export interface Resource {
  id: number;
  title: string;
  description: string;
  category: string;
  type: string;
  actionLabel: string;
  href: string;
}