# lib/api-zod/src/generated/types/siteSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and enquiry API for the Airlink Aviation website
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSummary {
  productCount: number;
  capabilityCount: number;
  supportLabel: string;
  qualityLabel: string;
}