# lib/api-zod/src/generated/types/specification.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public content and enquiry API for the Airlink Aviation website
 * OpenAPI spec version: 0.1.0
 */

export interface Specification {
  label: string;
  value: string;
}